requirements.txt
# 핵심 의존성
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
asyncpg>=0.29.0
redis[hiredis]>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiohttp>=3.9.0
python-dotenv>=1.0.0
cryptography>=42.0.0

# 데이터 처리 (Python 3.13 호환 버전)
pandas>=2.2.0
numpy>=1.26.0

# 로깅
structlog>=23.2.0

# 시간대 처리
pytz>=2023.3

# JSON 처리
orjson>=3.9.0

# 환경설정
python-decouple>=3.8